requirements.txt
diskcache==5.6.3
llama_cpp_python==0.2.11
numpy==2.1.2
typing_extensions==4.12.2
